refactor(menu): all to one - TWIG-179
