+++ src/ec/.storybook/preview-head.html
@@ -1,25 +1,26 @@
 <link
   rel="stylesheet"
-  href="https://cdn1.fpfis.tech.ec.europa.eu/ecl/v2.18.0/ec-preset-legacy-website/styles/ecl-ec-preset-legacy-website.css"
-  integrity="sha256-M8QkxM0cOcTtR2rJz0FisZ+h2VcNOq94RRkdUy2L1PU=
-  sha384-oNfpOC4fBTE0LKjMa0fYfBS/H0DwuxX7Le3ccSug2NEpX0J7mvl33yDPzWuXbexF
-  sha512-z3zt8Y1anogS+LDI3xgSkF1RvgiMI/Zx4ps5ViY4DLd6pU7vPc9EUu7POqxDrqQqxePIko5Bqx6hMY/r4wX0ew=="
+  href="https://cdn1.fpfis.tech.ec.europa.eu/ecl/v2.19.0/ec-preset-legacy-website/styles/ecl-ec-preset-legacy-website.css"
+  integrity="sha256-7xFVySXU4jHemYxqlXskxDVZCRh496Ro2jB/bH+NKsM=
+   sha384-Tm98Wqq/OPwfX5ZRq70yKi6Z8bkMMivs+GrGzML7/m8j0WGP8fwmaCwBxITsmtJX
+   sha512-P8SCGlp0oZUWH7oq/ig8HkY1SW+lL4ova0o+3LfuOAblbVI/cc1QYer73TrPMK52fZgudIyLjI9oIlY88Zu0Tg=="
   crossorigin="anonymous"
   media="screen"
 />
 <link
   rel="stylesheet"
-  href="https://cdn1.fpfis.tech.ec.europa.eu/ecl/v2.18.0/ec-preset-legacy-website/styles/ecl-ec-preset-legacy-website-print.css"
-  integrity="sha256-5RxVOf2+vbJs8qpSPoy1s+SPGxNx73s+en7uAUONgTM= sha384-2PEuso8GKtO5OiBExExUL+BwFwLeSdWcQmi0ZqK2UpMV+JXo1A7nKIXKDMSwNty/
-  sha512-kxI7L/CEEn0wkdT4EcPqeBwrxE6P2JHSyUGFWIPOa08DepgxQF4phBOGkq//yvZp5X8Uo6UbLAH0kq1oBXlBKA=="
+  href="https://cdn1.fpfis.tech.ec.europa.eu/ecl/v2.19.0/ec-preset-legacy-website/styles/ecl-ec-preset-legacy-website-print.css"
+  integrity="sha256-sxozpGW/Pph3xHC0krYHNikfeIEXUi7oq83pPDkHPcY=
+   sha384-Um+X2RSB2LRCCrRAwNo70F/I8JLRNd4y2BBLmSHsJW+o0xpdGnkYLZV6wOhnegcD
+   sha512-W/ZwcxBTW3MWHxecnAo+dYHFii3tqy+t7vKxdKCfC7BRAePFuP2YjwN0MKELml2MQygGXs5EHvTl6lcxv2A/LQ=="
   crossorigin="anonymous"
   media="print"
 />
 <script
-  src="https://cdn1.fpfis.tech.ec.europa.eu/ecl/v2.18.0/ec-preset-legacy-website/scripts/ecl-ec-preset-legacy-website.js"
-  integrity="sha256-tFoGwiJ6Ot4A+KqMvG1p/sbhuUu55hSf6oY6voZxNDc=
-  sha384-YbX88RWTt34cIaDiAB4WVvatoN7AQQp+9GSZBNXU9oYOpP5rHji1LoFMWvcAnggE
-  sha512-djHZSFPig+rrCx4nU7cZhOomfVWYkg/0iYNeO7yagdlqktPaMCBWkssMhFVzbTCKR2bPl1c1fZM+CxbjdfroEA=="
+  src="https://cdn1.fpfis.tech.ec.europa.eu/ecl/v2.19.0/ec-preset-legacy-website/scripts/ecl-ec-preset-legacy-website.js"
+  integrity="sha256-yY3/+W5rvzNqUy7kJTR4vLOr/PCoAxBGrntYG2G/5LE= 
+  sha384-P7RZC4NjjrZ8UIuy4WqJKOH4apMwAFxV9/kqgpfv/a7NA+CBZVgffgfqZcxga7XK 
+  sha512-zpmABANZmMIxr6m/9ujNEbZqTvk8NSL7kqRBPLzHqSjhGp0TAJUtID7SIeN5yFIgan8PzFfBLBK75LvK1N5+IA=="
   crossorigin="anonymous"
 ></script>
 <script src="https://unpkg.com/svg4everybody@2.1.9/dist/svg4everybody.min.js"></script>
