--- src/ec/packages/ec-component-menu-harmonised/demo/data--group2.js
@@ -1,4 +1,0 @@
-import specData from '@ecl/ec-specs-menu-harmonised/demo/data--group2';
-import adapter from '../adapter';
-
-export default adapter(specData);
