--- src/ec/packages/ec-component-menu-harmonised/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@ecl-twig/ec-component-menu-harmonised",
-  "author": "European Commission",
-  "license": "EUPL-1.1",
-  "version": "2.18.0",
-  "description": "ECL Twig - EC Menu Harmonised",
-  "dependencies": {
-    "@ecl-twig/ec-component-icon": "2.18.0",
-    "@ecl-twig/ec-component-link": "2.18.0"
-  },
-  "devDependencies": {
-    "@ecl/ec-component-menu-harmonised": "2.18.0",
-    "@ecl/ec-resources-icons": "2.18.0",
-    "@ecl/ec-specs-menu-harmonised": "2.18.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ec-europa/ecl-twig.git"
-  },
-  "bugs": {
-    "url": "https://github.com/ec-europa/ecl-twig/issues"
-  },
-  "homepage": "https://github.com/ec-europa/ecl-twig",
-  "keywords": [
-    "ecl",
-    "europa-component-library",
-    "design-system",
-    "twig"
-  ]
-}
