--- src/ec/packages/ec-component-menu-standardised/.npmignore
@@ -1,3 +1,0 @@
-__snapshots__
-demo
-*.js
