--- src/ec/packages/ec-component-menu-standardised/CHANGELOG.md
@@ -1,14 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [2.18.0](https://github.com/ec-europa/ecl-twig/compare/v2.17.3...v2.18.0) (2019-12-16)
-
-### Bug Fixes
-
-- **hide-not-ready-components:** Hiding site headers containing a menu - TWIG-161 ([#267](https://github.com/ec-europa/ecl-twig/issues/267)) ([8ab03f3](https://github.com/ec-europa/ecl-twig/commit/8ab03f31b8387251fa6a22056e1730ccf5b9fff9))
-
-### Features
-
-- **menus:** Adding components (standardised and harmonised) - Twig 162 ([#265](https://github.com/ec-europa/ecl-twig/issues/265)) ([53ab675](https://github.com/ec-europa/ecl-twig/commit/53ab67551dc74d17b99f108336ef641a0401673c))
