--- src/ec/packages/ec-component-menu-standardised/demo/data--en.js
@@ -1,4 +1,0 @@
-import specData from '@ecl/ec-specs-menu-standardised/demo/data--en';
-import adapter from '../adapter';
-
-export default adapter(specData);
