--- src/ec/packages/ec-component-menu-standardised/demo/data--fr.js
@@ -1,4 +1,0 @@
-import specData from '@ecl/ec-specs-menu-standardised/demo/data--fr';
-import adapter from '../adapter';
-
-export default adapter(specData);
