--- src/ec/packages/ec-component-menu-standardised/menu-standardised.test.js
@@ -1,50 +1,0 @@
-import { merge, renderTwigFileAsNode } from '@ecl-twig/test-utils';
-
-// Import data for tests
-import enData from './demo/data--en';
-import frData from './demo/data--fr';
-
-describe('EC - Menu Standardised', () => {
-  const template =
-    '@ecl-twig/ec-component-menu-standardised/ecl-menu-standardised.html.twig';
-  const render = params => renderTwigFileAsNode(template, params);
-
-  describe('Default', () => {
-    test('renders correctly', () => {
-      expect.assertions(1);
-
-      return expect(render(enData)).resolves.toMatchSnapshot();
-    });
-
-    test('renders correctly with extra class names', () => {
-      expect.assertions(1);
-
-      const withExtraClasses = merge(enData, {
-        extra_classes: 'custom-button custom-button--test',
-      });
-
-      return expect(render(withExtraClasses)).resolves.toMatchSnapshot();
-    });
-
-    test('renders correctly with extra attributes', () => {
-      expect.assertions(1);
-
-      const withExtraAttributes = merge(enData, {
-        extra_attributes: [
-          { name: 'data-test', value: 'data-test-value' },
-          { name: 'data-test-1', value: 'data-test-value-1' },
-        ],
-      });
-
-      return expect(render(withExtraAttributes)).resolves.toMatchSnapshot();
-    });
-  });
-
-  describe('Translated', () => {
-    test('renders correctly', () => {
-      expect.assertions(1);
-
-      return expect(render(frData)).resolves.toMatchSnapshot();
-    });
-  });
-});
