+++ src/ec/packages/ec-component-menu/adapter.js
