+++ src/ec/packages/ec-component-menu/demo/data--en.js
@@ -1,0 +1,4 @@
+import specData from '@ecl/ec-specs-menu/demo/data--en';
+import adapter from '../adapter';
+
+export default adapter(specData);
