+++ src/ec/packages/ec-component-menu/demo/data--fr.js
@@ -1,0 +1,4 @@
+import specData from '@ecl/ec-specs-menu/demo/data--fr';
+import adapter from '../adapter';
+
+export default adapter(specData);
