+++ src/ec/packages/ec-component-menu/demo/data--group1.js
@@ -1,4 +1,4 @@
-import specData from '@ecl/ec-specs-menu-harmonised/demo/data--group1';
+import specData from '@ecl/ec-specs-menu/demo/data--group1';
 import adapter from '../adapter';
 
 specData.group1 = true;
