+++ src/ec/packages/ec-component-menu/demo/data--group2.js
@@ -1,0 +1,4 @@
+import specData from '@ecl/ec-specs-menu/demo/data--group2';
+import adapter from '../adapter';
+
+export default adapter(specData);
