+++ src/ec/packages/ec-component-menu/ecl-menu.html.twig
@@ -2,6 +2,7 @@
 
 {#
   Parameters:
+  - "group": (string) (default: 'group1'): 'group1' or 'group2'
   - "icon_path": (string) (default: ''): Path to the icon sprite
   - "items": (associative array) (default: {}): The menu items - format: [
   {
@@ -44,7 +45,12 @@
 
 {# Internal properties #}
 
-{% set _css_class = 'ecl-menu-standardised' %}
+{% set _group = group|default('group1') %}
+{% set _css_class = 'ecl-menu' %}
+{% set _menu_item_class = 'ecl-menu__link' %}
+{% set _menu_item_attributes = [{ name: 'data-ecl-menu-link'}] %}
+{% set _menu_list_item_attributes = 'data-ecl-menu-item' %}
+{% set _css_class = _css_class ~ ' ' ~ _css_class ~ '--' ~ _group %}
 
 {% if extra_classes is defined and extra_classes is not empty %}
   {% set _css_class = _css_class ~ ' ' ~ extra_classes %}
@@ -62,18 +68,18 @@
 
 <nav
   class="{{ _css_class }}"
-  data-ecl-menu-standardised
-  data-ecl-auto-init="MenuStandardised"
+  data-ecl-menu
+  data-ecl-auto-init="Menu"
   aria-expanded="false"
   {{ _extra_attributes|raw }}
 >
   <div class="ecl-container">
     <a
-      class="ecl-link ecl-link--standalone ecl-menu-standardised__toggle"
+      class="ecl-link ecl-link--standalone ecl-menu__toggle"
       href={{ toggle_path }}
-      data-ecl-menu-standardised-toggle
+      data-ecl-menu-toggle
     >
-      <div class="ecl-menu-standardised__toggle-open">
+      <div class="ecl-menu__toggle-open">
         {% include '@ecl-twig/ec-component-icon/ecl-icon.html.twig' with {
           icon: {
             size: 's',
@@ -84,7 +90,7 @@
         } only %}
         {{- toggle_label_open -}}
       </div>
-      <div class="ecl-menu-standardised__toggle-close">
+      <div class="ecl-menu__toggle-close">
         {% include '@ecl-twig/ec-component-icon/ecl-icon.html.twig' with {
           icon: {
             size: 's',
@@ -96,15 +102,17 @@
         {{- toggle_label_close -}}
       </div>
     </a>
-    <div class="ecl-menu-standardised__site-name">{{ site_name }}</div>
+  {% if _group == 'group1' %}
+    <div class="ecl-menu__site-name">{{ site_name }}</div>
+  {% endif %}
     <ul
-      class="ecl-menu-standardised__list"
-      data-ecl-menu-standardised-list
+      class="ecl-menu__list"
+      data-ecl-menu-list
     >
   {% if items is defined and items is not empty and items is iterable %}
     {% for item in items %}
       {% if item.link is defined and item.link is not empty %}
-        {% include '@ecl-twig/ec-component-menu-standardised/ecl-menu-standardised-item.html.twig' with {
+        {% include '@ecl-twig/ec-component-menu/ecl-menu-item.html.twig' with {
           item: item,
           icon_path: icon_path,
         } only %}
