+++ src/ec/packages/ec-component-menu/menu.story.js.bckp
@@ -2,28 +2,26 @@ import merge from 'deepmerge';
 import { storiesOf } from '@storybook/html';
 import { withNotes } from '@ecl-twig/storybook-addon-notes';
 import withCode from '@ecl-twig/storybook-addon-code';
+import iconPath from '@ecl/ec-resources-icons/dist/sprites/icons.svg';
 
-import defaultSprite from '@ecl/ec-resources-icons/dist/sprites/icons.svg';
-import logo from '@ecl/ec-resources-logo/logo--en.svg';
-import siteHeaderHarmonised from './ecl-site-header-harmonised.html.twig';
+import dataEn from './demo/data--en';
+import dataFr from './demo/data--fr';
 import dataGroup1 from './demo/data--group1';
 import dataGroup2 from './demo/data--group2';
+
+import menuStandardised from './ecl-menu-harmonised.html.twig';
 import notes from './README.md';
 
-storiesOf('Components/Site Headers/Harmonised', module)
+storiesOf('Components/Menus', module)
   .addDecorator(withNotes)
   .addDecorator(withCode)
   .add(
     'group 1',
     () =>
-      siteHeaderHarmonised(
+      menuStandardised(
         merge(dataGroup1, {
-          logo: {
-            src: logo,
-          },
           group: 'group1',
-          logged: true,
-          icon_file_path: defaultSprite,
+          icon_path: iconPath,
         })
       ),
     {
@@ -33,13 +31,10 @@ storiesOf('Components/Site Headers/Harmonised', module)
   .add(
     'group 2',
     () =>
-      siteHeaderHarmonised(
+      menuStandardised(
         merge(dataGroup2, {
-          logo: {
-            src: logo,
-          },
           group: 'group2',
-          icon_file_path: defaultSprite,
+          icon_path: iconPath,
         })
       ),
     {
