+++ src/ec/packages/ec-component-menu/package.json
@@ -1,17 +1,17 @@
 {
-  "name": "@ecl-twig/ec-component-menu-standardised",
+  "name": "@ecl-twig/ec-component-menu",
   "author": "European Commission",
   "license": "EUPL-1.1",
-  "version": "2.18.0",
-  "description": "ECL Twig - EC Menu Standardised",
+  "version": "2.19.0",
+  "description": "ECL Twig - EC Menu Harmonised",
   "dependencies": {
     "@ecl-twig/ec-component-icon": "2.18.0",
     "@ecl-twig/ec-component-link": "2.18.0"
   },
   "devDependencies": {
-    "@ecl/ec-component-menu-standardised": "2.18.0",
-    "@ecl/ec-resources-icons": "2.18.0",
-    "@ecl/ec-specs-menu-standardised": "2.18.0"
+    "@ecl/ec-component-menu": "2.19.0",
+    "@ecl/ec-resources-icons": "2.19.0",
+    "@ecl/ec-specs-menu": "2.19.0"
   },
   "publishConfig": {
     "access": "public"
