+++ src/ec/packages/ec-component-site-header-harmonised/demo/data--group1.js
@@ -1,5 +1,5 @@
 import specs from '@ecl/ec-specs-site-header-harmonised/demo/data--group1';
-import dataMenu from '../../ec-component-menu-harmonised/demo/data--group1';
+import dataMenu from '../../ec-component-menu/demo/data--group1';
 import adapter from '../adapter';
 
 specs.menu = dataMenu;
