+++ src/ec/packages/ec-component-site-header-harmonised/demo/data--group2.js
@@ -1,5 +1,5 @@
 import specs from '@ecl/ec-specs-site-header-harmonised/demo/data--group2';
-import dataMenu from '../../ec-component-menu-harmonised/demo/data--group2';
+import dataMenu from '../../ec-component-menu/demo/data--group2';
 import adapter from '../adapter';
 
 specs.menu = dataMenu;
