+++ src/ec/packages/ec-component-site-header-standardised/demo/data--en.js
@@ -1,5 +1,5 @@
 import specs from '@ecl/ec-specs-site-header-standardised/demo/data--en';
-import dataMenu from '../../ec-component-menu-standardised/demo/data--en';
+import dataMenu from '../../ec-component-menu/demo/data--en';
 import adapter from '../adapter';
 
 specs.menu = dataMenu;
