+++ src/ec/packages/ec-component-site-header-standardised/demo/data--fr.js
@@ -1,5 +1,5 @@
 import specs from '@ecl/ec-specs-site-header-standardised/demo/data--fr';
-import dataMenu from '../../ec-component-menu-standardised/demo/data--fr';
+import dataMenu from '../../ec-component-menu/demo/data--fr';
 import adapter from '../adapter';
 
 specs.menu = dataMenu;
