+++ src/ec/packages/ec-components/package.json
@@ -2,7 +2,7 @@
   "name": "@ecl-twig/ec-components",
   "author": "European Commission",
   "license": "EUPL-1.1",
-  "version": "2.18.0",
+  "version": "2.19.0",
   "description": "ECL Twig - EC Components",
   "publishConfig": {
     "access": "public"
@@ -37,8 +37,7 @@
     "@ecl-twig/ec-component-language-list": "2.18.0",
     "@ecl-twig/ec-component-link": "2.18.0",
     "@ecl-twig/ec-component-media-container": "2.18.0",
-    "@ecl-twig/ec-component-menu-harmonised": "2.18.0",
-    "@ecl-twig/ec-component-menu-standardised": "2.18.0",
+    "@ecl-twig/ec-component-menu": "2.19.0",
     "@ecl-twig/ec-component-menu-legacy": "2.18.0",
     "@ecl-twig/ec-component-message": "2.18.0",
     "@ecl-twig/ec-component-ordered-list": "2.18.0",
